+++ src/core/document/lat_lon_point_distance_feature_query.rs
@@ -604,7 +604,7 @@ impl DistanceScorerIntersectVisitor {
 
 impl IntersectVisitor for DistanceScorerIntersectVisitor {
   fn grow(&mut self, count: usize) -> Result<()> {
-    self.result.grow(count as i32)?;
+    self.result.grow(count)?;
     Ok(())
   }
