+++ src/core/util/packed/bulk_operation_packed14.rs
@@ -31,7 +31,7 @@ impl Decoder for BulkOperationPacked14 {
     mut blocks_offset: usize,
     values: &mut [i64],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) {
     for _ in 0..iterations {
       let block0 = blocks[blocks_offset];
@@ -126,7 +126,7 @@ impl Decoder for BulkOperationPacked14 {
     mut blocks_offset: usize,
     values: &mut [i64],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) {
     for _ in 0..iterations {
       let byte0 = blocks[blocks_offset] as i64;
@@ -162,7 +162,7 @@ impl Decoder for BulkOperationPacked14 {
     mut blocks_offset: usize,
     values: &mut [i32],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) -> Result<()> {
     for _ in 0..iterations {
       let block0 = blocks[blocks_offset];
@@ -258,7 +258,7 @@ impl Decoder for BulkOperationPacked14 {
     mut blocks_offset: usize,
     values: &mut [i32],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) -> Result<()> {
     for _ in 0..iterations {
       let byte0 = blocks[blocks_offset] as i32;
