+++ src/core/util/packed/bulk_operation_packed2.rs
@@ -32,7 +32,7 @@ impl Decoder for BulkOperationPacked2 {
     mut blocks_offset: usize,
     values: &mut [i64],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) {
     for _ in 0..iterations {
       let block = blocks[blocks_offset];
@@ -82,7 +82,7 @@ impl Decoder for BulkOperationPacked2 {
     mut blocks_offset: usize,
     values: &mut [i64],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) {
     for _ in 0..iterations {
       let block = blocks[blocks_offset];
@@ -104,7 +104,7 @@ impl Decoder for BulkOperationPacked2 {
     mut blocks_offset: usize,
     values: &mut [i32],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) -> Result<()> {
     for _ in 0..iterations {
       let block = blocks[blocks_offset];
@@ -155,7 +155,7 @@ impl Decoder for BulkOperationPacked2 {
     mut blocks_offset: usize,
     values: &mut [i32],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) -> Result<()> {
     for _ in 0..iterations {
       let block = blocks[blocks_offset];
