+++ src/core/util/packed/bulk_operation_packed7.rs
@@ -32,7 +32,7 @@ impl Decoder for BulkOperationPacked7 {
     mut blocks_offset: usize,
     values: &mut [i64],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) {
     for _ in 0..iterations {
       let block0 = blocks[blocks_offset];
@@ -137,7 +137,7 @@ impl Decoder for BulkOperationPacked7 {
     mut blocks_offset: usize,
     values: &mut [i64],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) {
     for _ in 0..iterations {
       let byte0 = blocks[blocks_offset] as u64;
@@ -168,7 +168,7 @@ impl Decoder for BulkOperationPacked7 {
     mut blocks_offset: usize,
     values: &mut [i32],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) -> Result<()> {
     for _ in 0..iterations {
       let block0 = blocks[blocks_offset];
@@ -255,7 +255,7 @@ impl Decoder for BulkOperationPacked7 {
     mut blocks_offset: usize,
     values: &mut [i32],
     mut values_offset: usize,
-    iterations: i32,
+    iterations: usize,
   ) -> Result<()> {
     for _ in 0..iterations {
       let byte0 = blocks[blocks_offset] as i32;
